requirements.txt
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pywin32==306
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
